+++ imio/dms/mail/tests/permissions_base.py
@@ -65,7 +65,7 @@ class TestPermissionsBaseIncomingMail(TestPermissionsBase):
         intids = getUtility(IIntIds)
         params = {
             "title": "Courrier 10",
-            "mail_type": "courrier",
+            "mail_type": "certificat",
             "internal_reference_no": "E0010",
             "sender": [RelationValue(intids.getId(self.portal["contacts"]["jeancourant"]))],
             "treating_groups": self.portal["contacts"]["plonegroup-organization"]["direction-generale"]["grh"].UID(),
@@ -1247,7 +1247,7 @@ class TestPermissionsBaseIncomingEmail(TestPermissionsBase):
         intids = getUtility(IIntIds)
         params = {
             "title": "Courrier 10",
-            "mail_type": "email",
+            "mail_type": "autre",
             "internal_reference_no": "E0010",
             "sender": [RelationValue(intids.getId(self.portal["contacts"]["jeancourant"]))],
             "treating_groups": self.portal["contacts"]["plonegroup-organization"]["direction-generale"]["grh"].UID(),