+++ imio/dms/mail/tests/test_subscribers.py
@@ -67,7 +67,7 @@ class TestSubscribers(unittest.TestCase, ImioTestHelpers):
             "c1",
             **{
                 "sender": [RelationValue(self.intids.getId(self.portal.contacts["electrabel"]))],
-                "mail_type": u"courrier",
+                "mail_type": u"certificat",
                 "title": u"title",
             }
         )