+++ imio/dms/mail/tests/test_vocabularies.py
@@ -166,13 +166,10 @@ class TestVocabularies(unittest.TestCase, ImioTestHelpers):
         self.assertEquals(
             voc_list,
             [
-                (u"courrier", u"Courrier"),
-                (u"email", u"Email"),
-                (u"recommande", u"Recommandé"),
                 (u"certificat", u"Certificat médical"),
-                (u"fax", u"Fax"),
                 (u"retour-recommande", u"Retour recommandé"),
                 (u"facture", u"Facture"),
+                (u"autre", u"Autre"),
             ],
         )
         voc_list = [(t.value, t.title) for t in get_settings_vta_table("omail_send_modes", choose=True)]
@@ -199,13 +196,10 @@ class TestVocabularies(unittest.TestCase, ImioTestHelpers):
         self.assertListEqual(
             voc_list,
             [
-                (u"courrier", u"Courrier"),
-                (u"email", u"Email"),
-                (u"recommande", u"Recommandé"),
                 (u"certificat", u"Certificat médical"),
-                (u"fax", u"Fax"),
                 (u"retour-recommande", u"Retour recommandé"),
                 (u"facture", u"Facture"),
+                (u"autre", u"Autre"),
             ],
         )
 
@@ -214,7 +208,7 @@ class TestVocabularies(unittest.TestCase, ImioTestHelpers):
         voc_list = [t.value for t in voc_inst(self.imail)]
         self.assertListEqual(
             voc_list,
-            [None, u"courrier", u"email", u"recommande", u"certificat", u"fax", u"retour-recommande", u"facture"]
+            [None, u"certificat", u"retour-recommande", u"facture", u"autre"]
         )
         settings = getUtility(IRegistry).forInterface(IImioDmsMailConfig, False)
         mail_types = settings.mail_types
@@ -224,7 +218,7 @@ class TestVocabularies(unittest.TestCase, ImioTestHelpers):
         voc_list = [t.value for t in voc_inst(self.imail)]
         self.assertListEqual(
             voc_list,
-            [None, u"email", u"recommande", u"certificat", u"fax", u"retour-recommande", u"facture"]
+            [None, u"retour-recommande", u"facture", u"autre"]
         )
 
     def test_PloneGroupInterfacesVocabulary(self):