+++ CHANGES.rst
@@ -8,6 +8,8 @@ Changelog
   [sgeulette]
 - Added defer_security_update parameter on IMPreManagerValidation wf adaptation
   [sgeulette]
+- Updated incoming mail_types and send_modes examples.
+  [chris-adam]
 
 3.1.4 (2026-05-29)
 ------------------